[FIX] mail, test_mail: move a test into test_mail
Fine-tuning of 50f1adb467e, the test should be moved
into test_mail

closes odoo/odoo#47172
