+++ addons/test_mail/tests/__init__.py
@@ -18,4 +18,3 @@ from . import test_discuss
 from . import test_performance
 from . import test_res_users
 from . import test_odoobot
-from . import test_mail_activity
